+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to `vue3-tabs-component` will be documented in this file
 
 ## 1.3.5 - 2023-09-03
 - Refactor `Tab` & `Tabs` component prop definitions to TypeScript
+- Add `id` & `options.storageKey` props to `Tabs` component, which are used when remembering last opened tab
 
 ## 1.3.4 - 2023-07-31
 - Fix prop shape for all custom class props to allow for regular Vue class prop shape (string, array, object)